Uploading labs made up to 05.11.21
Lab_3
Make a programm that swaps rows and columns in matrix of double numbers until the highest number is in left upper corner.

Lab_4
Make an abstract class Boiler with a lot properties and make two inheritor classes.

Lab_5
Add custom exception class to the Lab_3 programm.
Added example of an exception class.